progress percentage now is working as well.
this feacher will show a percentage of how many task done from a to.
this percentage will count from task you done vs all task's;
for example: if we have 4 task's to done and we done just one of theme.
our algoritm will be look like this: (1 / 4) * 100
the abouve math sentence will equal to 25.

next feacher should be a graph percentage of doned task.
i dont have any idea to show a good graph of it.